Pipe: run and output but crashes on return to main
Loop is correct for pipes, I think. At least up until the end.
Problem seems to be that in main loop afterwards, readline gets
EOF, which leads to it returning NULL. Either it protects that and
exits, or it does not, and crashes.

Is it that the parent process is now trying to run on a STDIN that does
not exist?
Have we overwritten it, or not restored it with the final command?
Alternatively: should we catch EOF and work with it? (HOW?)