CD: add copy_envp function to reduce duplication
We have to copy an envp to a temporary location and back again.
This was done inline, using space and variables that put us over
the Norm in ms_export_cd.

This change has been tested a bit and should be followed with the
removal of dead code and spare variables.