CD: Norm-compliant CD with no obvious memleaks
Completed the process of removing dead code and unused variables.
This file now only needs the header added, other minor changes are
nice-to-haves I think.

The created variables for oldpwd and newpwd do not appear to cause
memleaks.
I think they are freed when the env of the shell is.