Exec: Fix "must exit twice" after failed command
If the last command in a pipe fails to execve (because it does
not exist or bad permissions) then we were left running in the
child process.
It looked the same but meant that "exit" took the user only out of
that child process, returning them to the identical-looking parent.
It looked like exit had done nothing.

By calling an exit at the end of run_command, we are guaranteed to exit
the child process on success (execve overwrites child process) or
failure (exit_failed_pipe is reached).