Parse: add grab_control_seq to handle >test.txt
Bash does not require a space between control characters and the
other parts of the command. This new function allows to treat
ls >test and ls > test as equivalent by splitting control sequences
into their own parameter.
Is only implemented for raw_parameter so far, and the memory behaviour
still needs to be tested.