PATHS: fix so non-forking builtins only run once
Without this check it was possible for a builtin to return
and immediately be sent to run_final_command().
Had to remove an error display to fit norm, but we can  live without
it, very rare case.
Example: cd non_existent_dir
lead to a double error message being printed.
Otherwise was hard to spot, not sure if there were other effects.